# crates/selene-server/openapi.yaml
openapi: '3.1.0'
info:
  title: Selene API
  version: 0.2.0
  description: Lightweight in-memory property graph runtime for IoT, smart buildings, and any domain requiring a living graph of connected entities with real-time state.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /nodes:
    get:
      operationId: listNodes
      summary: List nodes with optional label filter
      tags: [Nodes]
      parameters:
        - name: label
          in: query
          schema:
            type: string
          description: Filter by node label
          example: sensor
        - name: limit
          in: query
          schema:
            type: integer
            format: int64
            default: 1000
            maximum: 10000
          description: Maximum number of nodes to return
        - name: offset
          in: query
          schema:
            type: integer
            format: int64
            default: 0
          description: Number of nodes to skip
      responses:
        '200':
          description: Paginated list of nodes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createNode
      summary: Create a new node
      tags: [Nodes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNodeRequest'
            example:
              labels: [sensor]
              properties:
                name: "Temperature Sensor 1"
                floor: 3
                unit: "celsius"
              parent_id: 1
      responses:
        '201':
          description: Node created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/ReadOnly'
        '422':
          $ref: '#/components/responses/SchemaViolation'
  /nodes/{id}:
    get:
      operationId: getNode
      summary: Get a node by ID
      tags: [Nodes]
      parameters:
        - $ref: '#/components/parameters/NodeId'
      responses:
        '200':
          description: Node details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: modifyNode
      summary: Modify a node (set/remove properties and labels)
      tags: [Nodes]
      parameters:
        - $ref: '#/components/parameters/NodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyNodeRequest'
            example:
              set_properties:
                temp: 72.5
                status: "active"
              remove_properties: [old_field]
              add_labels: [active]
              remove_labels: [inactive]
      responses:
        '200':
          description: Updated node
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/ReadOnly'
        '422':
          $ref: '#/components/responses/SchemaViolation'
    delete:
      operationId: deleteNode
      summary: Delete a node and its connected edges
      tags: [Nodes]
      parameters:
        - $ref: '#/components/parameters/NodeId'
      responses:
        '204':
          description: Node deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/ReadOnly'
  /nodes/{id}/edges:
    get:
      operationId: listNodeEdges
      summary: List all edges connected to a node
      tags: [Nodes]
      parameters:
        - $ref: '#/components/parameters/NodeId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 1000
            maximum: 10000
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Edges connected to the node
          content:
            application/json:
              schema:
                type: object
                required: [node_id, edges, total]
                properties:
                  node_id:
                    type: integer
                    format: uint64
                  edges:
                    type: array
                    items:
                      $ref: '#/components/schemas/Edge'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Edges ─────────────────────────────────────────────────────────
  /edges:
    get:
      operationId: listEdges
      summary: List edges with optional label filter
      tags: [Edges]
      parameters:
        - name: label
          in: query
          schema:
            type: string
          description: Filter by edge label
          example: contains
        - name: limit
          in: query
          schema:
            type: integer
            format: int64
            default: 1000
            maximum: 10000
        - name: offset
          in: query
          schema:
            type: integer
            format: int64
            default: 0
      responses:
        '200':
          description: Paginated list of edges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgeList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createEdge
      summary: Create a new edge between two nodes
      tags: [Edges]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEdgeRequest'
            example:
              source: 1
              target: 2
              label: contains
              properties:
                priority: 1
      responses:
        '201':
          description: Edge created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edge'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/ReadOnly'
        '422':
          $ref: '#/components/responses/SchemaViolation'
  /edges/{id}:
    get:
      operationId: getEdge
      summary: Get an edge by ID
      tags: [Edges]
      parameters:
        - $ref: '#/components/parameters/EdgeId'
      responses:
        '200':
          description: Edge details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edge'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: modifyEdge
      summary: Modify edge properties
      tags: [Edges]
      parameters:
        - $ref: '#/components/parameters/EdgeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyEdgeRequest'
            example:
              set_properties:
                weight: 0.95
              remove_properties: [old_field]
      responses:
        '200':
          description: Updated edge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edge'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/ReadOnly'
    delete:
      operationId: deleteEdge
      summary: Delete an edge
      tags: [Edges]
      parameters:
        - $ref: '#/components/parameters/EdgeId'
      responses:
        '204':
          description: Edge deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/ReadOnly'

  # ── Time-Series ───────────────────────────────────────────────────
  /ts/samples:
    post:
      operationId: writeTsSamples
      summary: Write time-series samples
      tags: [TimeSeries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TsWriteRequest'
            example:
              samples:
                - entity_id: 42
                  property: temperature
                  timestamp_nanos: 1711699200000000000
                  value: 72.5
                - entity_id: 42
                  property: humidity
                  timestamp_nanos: 1711699200000000000
                  value: 45.0
      responses:
        '200':
          description: Samples written
          content:
            application/json:
              schema:
                type: object
                required: [written]
                properties:
                  written:
                    type: integer
                    description: Number of samples written
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /ts/write:
    post:
      operationId: writeTsSamplesDeprecated
      summary: Write time-series samples (deprecated alias for /ts/samples)
      deprecated: true
      tags: [TimeSeries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TsWriteRequest'
      responses:
        '200':
          description: Samples written
          content:
            application/json:
              schema:
                type: object
                required: [written]
                properties:
                  written:
                    type: integer
  /ts/{entity_id}/{property}:
    get:
      operationId: queryTimeSeries
      summary: Query time-series samples for an entity property
      tags: [TimeSeries]
      parameters:
        - name: entity_id
          in: path
          required: true
          schema:
            type: integer
            format: uint64
          description: Node ID
        - name: property
          in: path
          required: true
          schema:
            type: string
          description: Property name
          example: temperature
        - name: start
          in: query
          schema:
            type: integer
            format: int64
            default: 0
          description: Start timestamp (nanoseconds since Unix epoch)
        - name: end
          in: query
          schema:
            type: integer
            format: int64
          description: End timestamp (nanoseconds since Unix epoch)
        - name: limit
          in: query
          schema:
            type: integer
          description: Maximum number of samples to return
      responses:
        '200':
          description: Time-series samples
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TsSample'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── GQL ───────────────────────────────────────────────────────────
  /gql:
    post:
      operationId: executeGql
      summary: Execute a GQL query or mutation
      description: |
        Primary query interface. Supports ISO GQL (ISO 39075) queries, mutations,
        DDL statements, and CALL procedures. Returns GQLSTATUS codes in the response.
        Read-only queries bypass the mutation batcher for lower latency.
      tags: [GQL]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GqlRequest'
            examples:
              simple_query:
                summary: Match sensors by temperature
                value:
                  query: "MATCH (s:sensor) FILTER s.temp > 72 RETURN s.name AS name, s.temp AS temp"
              parameterized:
                summary: Parameterized query
                value:
                  query: "MATCH (n:equipment) FILTER n.floor = $floor RETURN n"
                  parameters:
                    floor: 3
              explain:
                summary: Explain query plan
                value:
                  query: "MATCH (a:sensor)-[:monitors]->(b:equipment) RETURN a, b"
                  explain: true
              mutation:
                summary: Insert a node
                value:
                  query: "INSERT (s:sensor {name: 'Temp-01', unit: 'celsius'})"
      responses:
        '200':
          description: GQL execution result (includes GQLSTATUS even on query errors)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GqlResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Graph ─────────────────────────────────────────────────────────
  /graph/slice:
    post:
      operationId: graphSlice
      summary: Retrieve a graph slice (full, by labels, or containment tree)
      tags: [Graph]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphSliceRequest'
            examples:
              full:
                summary: Full graph slice
                value:
                  slice_type: full
                  limit: 100
                  offset: 0
              labels:
                summary: Filter by labels
                value:
                  slice_type: labels
                  labels: [sensor, equipment]
              containment:
                summary: Containment tree from root
                value:
                  slice_type: containment
                  root_id: 1
                  max_depth: 5
      responses:
        '200':
          description: Graph slice with nodes and edges
          content:
            application/json:
              schema:
                type: object
                required: [nodes, edges]
                properties:
                  nodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Node'
                  edges:
                    type: array
                    items:
                      $ref: '#/components/schemas/Edge'
                  total_nodes:
                    type: integer
                    description: Total node count (present for full/labels slices)
                  total_edges:
                    type: integer
                    description: Total edge count (present for full/labels slices)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /graph/stats:
    get:
      operationId: graphStats
      summary: Graph statistics (node/edge counts by label)
      tags: [Graph]
      responses:
        '200':
          description: Graph statistics
          content:
            application/json:
              schema:
                type: object
                required: [node_count, edge_count, node_labels, edge_labels]
                properties:
                  node_count:
                    type: integer
                    format: uint64
                  edge_count:
                    type: integer
                    format: uint64
                  node_labels:
                    type: object
                    additionalProperties:
                      type: integer
                      format: uint64
                    example:
                      sensor: 150
                      equipment: 40
                      building: 1
                  edge_labels:
                    type: object
                    additionalProperties:
                      type: integer
                      format: uint64
                    example:
                      contains: 190
                      monitors: 75
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /schemas:
    get:
      operationId: listSchemas
      summary: List all registered node and edge schemas
      tags: [Schemas]
      responses:
        '200':
          description: All registered schemas
          content:
            application/json:
              schema:
                type: object
                required: [node_schemas, edge_schemas]
                properties:
                  node_schemas:
                    type: array
                    items:
                      $ref: '#/components/schemas/NodeSchema'
                  edge_schemas:
                    type: array
                    items:
                      $ref: '#/components/schemas/EdgeSchema'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /export/csv:
    get:
      operationId: csvExportDeprecated
      summary: Export CSV (deprecated alias for GET /graph/csv)
      deprecated: true
      tags: [CSV]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [nodes, edges]
            default: nodes
        - name: label
          in: query
          schema:
            type: string
      responses:
        '200':
          description: CSV data
          content:
            text/csv:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── RDF ───────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Format: identity:secret"
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ReadOnly:
      description: Mutation rejected on read-only replica
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "read-only replica"
...
  schemas:
...
    NodeList:
      type: object
      required: [nodes, total]
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/Node'
        total:
          type: integer
          description: Total count (before pagination)
...
    CreateNodeRequest:
      type: object
      required: [labels]
      properties:
        labels:
          type: array
          items:
            type: string
          minItems: 1
          example: [sensor]
        properties:
          type: object
          additionalProperties: {}
          description: Plain JSON key-value pairs. Nested objects are stored as JSON strings.
          default: {}
        parent_id:
          type: integer
          format: uint64
          description: If set, auto-creates a "contains" edge from this parent node
...